requirements.txt
Pillow>=9.5
pycocotools>=2.0.7
